+++ src/main/java/net/sf/jabref/gui/exporter/ExportCustomizationDialog.java
@@ -40,9 +40,9 @@ import javax.swing.table.TableColumnModel;
 import net.sf.jabref.Globals;
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.gui.help.HelpAction;
-import net.sf.jabref.gui.help.HelpFile;
 import net.sf.jabref.gui.keyboard.KeyBinding;
 import net.sf.jabref.gui.util.FocusRequester;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.l10n.Localization;
 
 import ca.odell.glazedlists.gui.TableFormat;