+++ src/main/java/net/sf/jabref/importer/fetcher/SpringerFetcher.java
@@ -19,10 +19,10 @@ package net.sf.jabref.importer.fetcher;
     import javax.swing.JOptionPane;
     import javax.swing.JPanel;
 
-    import net.sf.jabref.gui.help.HelpFile;
     import net.sf.jabref.importer.ImportInspector;
     import net.sf.jabref.importer.OutputPrinter;
     import net.sf.jabref.importer.fileformat.JSONEntryParser;
+    import net.sf.jabref.logic.help.HelpFile;
     import net.sf.jabref.logic.l10n.Localization;
     import net.sf.jabref.model.entry.BibEntry;